+++ packages/lumx-react/src/components/link/Link.test.tsx
@@ -1,75 +1,102 @@
-import { ColorPalette, ColorVariant, Icon, Typography } from '@lumx/react';
-import { commonTestsSuiteRTL } from '@lumx/react/testing/utils';
-import { getByClassName, queryAllByClassName, queryByClassName } from '@lumx/react/testing/utils/queries';
+import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+
+import BaseLinkTests, { setup } from '@lumx/core/js/components/Link/Tests';
+import { LinkProps } from '@lumx/core/js/components/Link';
+import { commonTestsSuiteRTL, SetupRenderOptions } from '@lumx/react/testing/utils';
+import { Icon } from '@lumx/react';
 import { mdiCheck, mdiPlus } from '@lumx/icons';
-import { Link, LinkProps } from './Link';
+
+import { queryAllByClassName } from '@lumx/react/testing/utils/queries';
+import { Link } from './Link';
 
 const CLASSNAME = Link.className as string;
 
-/**
- * Mounts the component and returns common DOM elements / data needed in multiple tests further down.
- */
-const setup = (props: LinkProps = {}) => {
-    render(<Link {...props} />);
-    const link = getByClassName(document.body, CLASSNAME);
-    const rightIcon = queryByClassName(link, `${CLASSNAME}__right-icon`);
-    const leftIcon = queryByClassName(link, `${CLASSNAME}__left-icon`);
-    return { props, link, rightIcon, leftIcon };
-};
-
 describe(`<${Link.displayName}>`, () => {
-    describe('Props', () => {
-        it('should render link', () => {
-            const name = 'Link';
-            const { props, link, rightIcon, leftIcon } = setup({ href: 'https://example.com', children: name });
+    const renderLink = (props: LinkProps, options?: SetupRenderOptions) => {
+        // Map core props to React props (label -> children)
+        const { label, ...restProps } = props;
+        return render(<Link {...(restProps as any)}>{label}</Link>, options);
+    };
 
-            expect(link).toBe(screen.queryByRole('link', { name }));
-            expect(link).toHaveAttribute('href', props.href);
-            expect(link.className).toMatchInlineSnapshot('"lumx-link"');
+    BaseLinkTests({ render: renderLink, screen });
 
-            expect(rightIcon).not.toBeInTheDocument();
-            expect(leftIcon).not.toBeInTheDocument();
+    const setupLink = (props: Partial<LinkProps> = {}, options: SetupRenderOptions = {}) =>
+        setup(props, { ...options, render: renderLink, screen });
+
+    describe('React', () => {
+        it('should forward ref to the root element', () => {
+            const ref = React.createRef<HTMLAnchorElement>();
+            render(<Link href="https://example.com" ref={ref} />);
+            expect(ref.current).toBeInstanceOf(HTMLAnchorElement);
         });
 
-        it('should render color & color variant', () => {
-            const { link } = setup({
-                href: 'https://google.com',
-                color: ColorPalette.primary,
-                colorVariant: ColorVariant.D1,
-            });
-            expect(link.className).toBe('lumx-link lumx-link--color-primary lumx-link--color-variant-D1');
+        it('should render as button and forward ref', () => {
+            const ref = React.createRef<HTMLButtonElement>();
+            // eslint-disable-next-line jsx-a11y/anchor-is-valid
+            render(<Link ref={ref}>Link</Link>);
+            expect(ref.current).toBeInstanceOf(HTMLButtonElement);
         });
 
-        it('should render typography', () => {
-            const { link } = setup({ href: 'https://google.com', typography: Typography.title });
-            expect(link.className).toBe('lumx-link lumx-link--has-typography lumx-typography-title');
-        });
-
-        it('should render a button', () => {
-            const name = 'Link';
+        it('should be disabled with isDisabled', async () => {
             const onClick = vi.fn();
-            const { link } = setup({ onClick, children: name });
-            expect(link).toBe(screen.queryByRole('button', { name }));
+            const { container } = render(
+                <Link href="/test" isDisabled onClick={onClick}>
+                    Label
+                </Link>,
+            );
+            const link = container.querySelector(`.${CLASSNAME}`);
+
+            expect(link).toHaveAttribute('disabled');
+
+            // Should not trigger onClick.
+            if (link) await userEvent.click(link);
+            expect(onClick).not.toHaveBeenCalled();
         });
 
-        it('should render as a different element', () => {
-            setup({ linkAs: 'span', children: 'Link' });
-            expect(screen.getByText('Link').closest('span')).toBeInTheDocument();
+        it('should be disabled link with isDisabled and href', async () => {
+            const onClick = vi.fn();
+            const { container } = render(
+                <Link isDisabled href="https://example.com" onClick={onClick}>
+                    Label
+                </Link>,
+            );
+            const link = container.querySelector(`.${CLASSNAME}`);
+
+            expect(link).toHaveAttribute('aria-disabled', 'true');
+            expect(link).toHaveAttribute('tabindex', '-1');
+
+            // Should not trigger onClick.
+            if (link) await userEvent.click(link);
+            expect(onClick).not.toHaveBeenCalled();
         });
 
-        it('should forward target prop to anchor', () => {
-            const { link } = setup({ href: 'https://example.com', target: '_blank', children: 'Link' });
-            expect(link).toHaveAttribute('target', '_blank');
+        it('should be disabled with aria-disabled', async () => {
+            const onClick = vi.fn();
+            const { container } = render(
+                <Link href="/test" aria-disabled onClick={onClick}>
+                    Label
+                </Link>,
+            );
+            const link = container.querySelector(`.${CLASSNAME}`);
+
+            expect(link).toHaveAttribute('aria-disabled', 'true');
+
+            // Should not trigger onClick.
+            if (link) await userEvent.click(link);
+            expect(onClick).not.toHaveBeenCalled();
         });
 
-        it('should render with icons', () => {
-            const { link } = setup({
-                leftIcon: mdiCheck,
-                children: ['Link', <Icon key="icon" icon={mdiCheck} />, 'with icons'],
-                rightIcon: mdiPlus,
-            });
+        it('should render with icons wrapped with spaces', () => {
+            const { container } = render(
+                <Link href="/test" leftIcon={mdiCheck} rightIcon={mdiPlus}>
+                    Link
+                    <Icon icon={mdiCheck} />
+                    with icons
+                </Link>,
+            );
+            const link = container.querySelector(`.${CLASSNAME}`) as HTMLElement;
             const icons = queryAllByClassName(link, Icon.className as string);
             expect(icons).toHaveLength(3);
 
@@ -81,54 +108,7 @@ describe(`<${Link.displayName}>`, () => {
         });
     });
 
-    describe('Disabled state', () => {
-        it('should render disabled button', async () => {
-            const onClick = vi.fn();
-            const { link } = setup({ children: 'Label', isDisabled: true, onClick });
-            expect(link).toHaveAttribute('disabled');
-            await userEvent.click(link);
-            expect(onClick).not.toHaveBeenCalled();
-        });
-
-        it('should render disabled link', async () => {
-            const onClick = vi.fn();
-            const { link } = setup({ children: 'Label', isDisabled: true, href: 'https://example.com', onClick });
-            expect(screen.queryByRole('link')).toBeInTheDocument();
-            expect(link).toHaveAttribute('aria-disabled');
-            // Simulate standard disabled state (not focusable)
-            expect(link).toHaveAttribute('tabindex', '-1');
-            await userEvent.click(link);
-            expect(onClick).not.toHaveBeenCalled();
-        });
-
-        it('should render aria-disabled button', async () => {
-            const onClick = vi.fn();
-            const { link } = setup({ children: 'Label', 'aria-disabled': true, onClick });
-            expect(screen.queryByRole('button')).toBeInTheDocument();
-            expect(link).toHaveAttribute('aria-disabled', 'true');
-            expect(link).not.toHaveAttribute('tabindex');
-            await userEvent.click(link);
-            expect(onClick).not.toHaveBeenCalled();
-        });
-
-        it('should render aria-disabled link', async () => {
-            const onClick = vi.fn();
-            const { link } = setup({
-                children: 'Label',
-                'aria-disabled': true,
-                href: 'https://example.com',
-                onClick,
-            });
-            expect(link).toHaveAccessibleName('Label');
-            expect(screen.queryByRole('link')).toBeInTheDocument();
-            expect(link).toHaveAttribute('aria-disabled', 'true');
-            await userEvent.click(link);
-            expect(onClick).not.toHaveBeenCalled();
-        });
-    });
-
-    // Common tests suite.
-    commonTestsSuiteRTL(setup, {
+    commonTestsSuiteRTL(setupLink, {
         baseClassName: CLASSNAME,
         forwardAttributes: 'link',
         forwardClassName: 'link',