+++ packages/lumx-vue/src/index.ts
@@ -11,6 +11,7 @@ export * from './components/heading';
 export * from './components/icon';
 export * from './components/input-helper';
 export * from './components/input-label';
+export * from './components/link';
 export * from './components/message';
 export * from './components/radio-button';
 export * from './components/skeleton';