+++ packages/site-demo/content/product/components/link/vue/demo3.vue
@@ -1,0 +1,14 @@
+<template>
+    <Link href="#" color="dark">
+        Dark color link
+    </Link>
+    <div class="lumx-spacing-padding lumx-color-background-dark-N">
+        <Link href="#" color="light">
+            Light color link
+        </Link>
+    </div>
+</template>
+
+<script setup lang="ts">
+import { Link } from '@lumx/vue';
+</script>